pyproject.toml
[project]
name = "reqana"
version = "0.1.0"
description = "Consulting requirement intake from files and voice, organized into requirement documents."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openai>=1.93.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
